+++ routes/api.php
@@ -18,6 +18,8 @@ use Illuminate\Support\Facades\Route;
 | group reading config('access.route_prefix') and config('access.middleware').
 */
 
+Route::post('modules/bulk', [ModuleController::class, 'bulkStore']);
+Route::delete('modules/bulk', [ModuleController::class, 'bulkDestroy']);
 Route::apiResource('modules', ModuleController::class);
 
 Route::get('roles', [RoleController::class, 'index']);
@@ -26,6 +28,7 @@ Route::get('roles/{role}', [RoleController::class, 'show']);
 Route::put('roles/{role}', [RoleController::class, 'update']);
 Route::delete('roles/{role}', [RoleController::class, 'destroy']);
 Route::post('roles/{role}/clone', [RoleController::class, 'clone']);
+Route::post('roles/{role}/users/bulk', [RoleController::class, 'bulkAssignUsers']);
 Route::put('roles/{role}/modules', [RoleController::class, 'syncModules']);
 Route::get('roles/{role}/permission-matrix', [RoleController::class, 'permissionMatrix']);